Cancel an in-progress operation

+++ Banyan/Media+Transfer.h
@@ -14,14 +14,14 @@
 - (void)uploadWithSuccess:(void (^)())success failure:(void (^)(NSError *error))failure;
 - (void) deleteWitSuccess:(void (^)(void))successBlock failure:(void (^)(NSError *error))errorBlock;
 - (void) getImageForMediaWithSuccess:(void (^)(UIImage *image))success
-                            progress:(void (^)(NSUInteger receivedSize, long long expectedSize))progress
+                            progress:(void (^)(NSInteger receivedSize, NSInteger expectedSize))progress
                              failure:(void (^)(NSError *error))failure
                     includeThumbnail:(BOOL) includeThumbnail;
 - (void) getImageWithContentMode:(UIViewContentMode)contentMode
                           bounds:(CGSize)size
             interpolationQuality:(CGInterpolationQuality)quality
              forMediaWithSuccess:(void (^)(UIImage *image))success
-                        progress:(void (^)(NSUInteger receivedSize, long long expectedSize))progress
+                        progress:(void (^)(NSInteger receivedSize, NSInteger expectedSize))progress
                          failure:(void (^)(NSError *error))failure
                 includeThumbnail:(BOOL) includeThumbnail;
